Fix all compilation errors and CI issues for CMake build.
- Added Eigen3 dependency and linked it in CMakeLists.txt.
- Fixed shader string buffer overflows by increasing sizes and using snprintf.
- Resolved giflib 5.x API compatibility issues.
- Fixed libpng usage and syntax errors in pngSave.cpp.
- Ensured GLU headers are included where needed.
- Fixed missing return values and default arguments in source files.
- Updated CI workflow to install all required dependencies (Ubuntu and Windows).
- Fixed printf format string warnings for size_t.

+++ src/Mol.cpp
@@ -301,7 +301,7 @@ void Mol::FindBonds(){
     
   
   if (showprogress) EndProgress();
-  printf("%d bonds detected!\n",bond.size() );
+  printf("%zd bonds detected!\n", (size_t)bond.size() );
 }
 
 void Mol::SetSpaceFill(){
